Revert timer
